Move provider-specific options and their UI to the provider itself
