+++ picard/ui/options/cover.py
@@ -30,194 +30,6 @@ from picard.util import webbrowser2
 from picard.coverart.providers import cover_art_providers, is_provider_enabled
 
 
-class CAATypesSelectorDialog(QtGui.QDialog):
-    _columns = 4
-
-    def __init__(self, parent=None, types=[]):
-        super(CAATypesSelectorDialog, self).__init__(parent)
-
-        self.setWindowTitle(_("Cover art types"))
-        self._items = {}
-        self.layout = QtGui.QVBoxLayout(self)
-
-        grid = QtGui.QWidget()
-        gridlayout = QtGui.QGridLayout()
-        grid.setLayout(gridlayout)
-
-        rows = len(CAA_TYPES) // self._columns + 1
-        positions = [(i, j) for i in range(rows) for j in range(self._columns)]
-
-        for position, caa_type in zip(positions, CAA_TYPES):
-            name = caa_type["name"]
-            text = translate_caa_type(name)
-            item = QtGui.QCheckBox(text)
-            item.setChecked(name in types)
-            self._items[item] = caa_type
-            gridlayout.addWidget(item, *position)
-
-        self.layout.addWidget(grid)
-
-        self.buttonbox = QtGui.QDialogButtonBox(self)
-        self.buttonbox.setOrientation(QtCore.Qt.Horizontal)
-        self.buttonbox.addButton(
-            StandardButton(StandardButton.OK), QtGui.QDialogButtonBox.AcceptRole)
-        self.buttonbox.addButton(StandardButton(StandardButton.CANCEL),
-                                 QtGui.QDialogButtonBox.RejectRole)
-        self.buttonbox.addButton(
-            StandardButton(StandardButton.HELP), QtGui.QDialogButtonBox.HelpRole)
-        self.buttonbox.accepted.connect(self.accept)
-        self.buttonbox.rejected.connect(self.reject)
-        self.buttonbox.helpRequested.connect(self.help)
-
-        extrabuttons = [
-            (N_("Chec&k all"), self.checkall),
-            (N_("&Uncheck all"), self.uncheckall),
-        ]
-        for label, callback in extrabuttons:
-            button = QtGui.QPushButton(_(label))
-            button.setSizePolicy(
-                QtGui.QSizePolicy.Fixed, QtGui.QSizePolicy.Expanding)
-            self.buttonbox.addButton(button, QtGui.QDialogButtonBox.ActionRole)
-            button.clicked.connect(callback)
-
-        self.layout.addWidget(self.buttonbox)
-
-        self.buttonbox.accepted.connect(self.accept)
-        self.buttonbox.rejected.connect(self.reject)
-
-    def help(self):
-        webbrowser2.goto('doc_cover_art_types')
-
-    def uncheckall(self):
-        self._set_checked_all(False)
-
-    def checkall(self):
-        self._set_checked_all(True)
-
-    def _set_checked_all(self, value):
-        for item in self._items.keys():
-            item.setChecked(value)
-
-    def get_selected_types(self):
-        types = []
-        for item, typ in self._items.iteritems():
-            if item.isChecked():
-                types.append(typ['name'])
-        if not types:
-            return [u'front']
-        return types
-
-    @staticmethod
-    def run(parent=None, types=[]):
-        dialog = CAATypesSelectorDialog(parent, types)
-        result = dialog.exec_()
-        return (dialog.get_selected_types(), result == QtGui.QDialog.Accepted)
-
-
-class ProviderOptions(QtGui.QWidget):
-
-    """
-        Abstract class for provider's options
-    """
-
-    options = []
-
-    _options_ui = None
-
-    def __init__(self, options_page, parent=None):
-        super(ProviderOptions, self).__init__(parent)
-        self.options_page = options_page
-        if callable(self._options_ui):
-            self.ui = self._options_ui()
-            self.ui.setupUi(self)
-
-    def load(self):
-        pass
-
-    def save(self):
-        pass
-
-
-from picard.ui.ui_provider_options_caa import Ui_CaaOptions
-class ProviderOptionsCaa(ProviderOptions):
-    """
-        Options for Cover Art Archive cover art provider
-    """
-
-    options = [
-        config.BoolOption("setting", "caa_approved_only", False),
-        config.BoolOption("setting", "caa_image_type_as_filename", False),
-        config.IntOption("setting", "caa_image_size", 1),
-        config.ListOption("setting", "caa_image_types", [u"front"]),
-        config.BoolOption("setting", "caa_restrict_image_types", True),
-    ]
-
-    _options_ui = Ui_CaaOptions
-
-    def __init__(self, options_page, parent=None):
-        super(ProviderOptionsCaa, self).__init__(options_page, parent)
-        self.ui.restrict_images_types.clicked.connect(self.update_caa_types)
-        self.ui.select_caa_types.clicked.connect(self.select_caa_types)
-
-    def load(self):
-        self.ui.cb_image_size.setCurrentIndex(config.setting["caa_image_size"])
-        self.ui.cb_approved_only.setChecked(config.setting["caa_approved_only"])
-        self.ui.cb_type_as_filename.setChecked(config.setting["caa_image_type_as_filename"])
-        self.ui.restrict_images_types.setChecked(
-            config.setting["caa_restrict_image_types"])
-        self.update_caa_types()
-
-    def save(self):
-        config.setting["caa_image_size"] =\
-            self.ui.cb_image_size.currentIndex()
-        config.setting["caa_approved_only"] =\
-            self.ui.cb_approved_only.isChecked()
-        config.setting["caa_image_type_as_filename"] = \
-            self.ui.cb_type_as_filename.isChecked()
-        config.setting["caa_restrict_image_types"] = \
-            self.ui.restrict_images_types.isChecked()
-
-    def update_caa_types(self):
-        enabled = self.ui.restrict_images_types.isChecked()
-        self.ui.select_caa_types.setEnabled(enabled)
-
-    def select_caa_types(self):
-        (types, ok) = CAATypesSelectorDialog.run(
-            self.options_page, config.setting["caa_image_types"])
-        if ok:
-            config.setting["caa_image_types"] = types
-
-
-from picard.ui.ui_provider_options_local import Ui_LocalOptions
-class ProviderOptionsLocal(ProviderOptions):
-    """
-        Options for Local Files cover art provider
-    """
-
-    _DEFAULT_LOCAL_COVER_ART_REGEX = '^(?:cover|folder|albumart)(.*)\.(?:jpe?g|png|gif|tiff?)$'
-
-    options = [
-        config.TextOption("setting", "local_cover_regex",
-                          _DEFAULT_LOCAL_COVER_ART_REGEX),
-    ]
-
-    _options_ui = Ui_LocalOptions
-
-    def __init__(self, options_page, parent=None):
-        super(ProviderOptionsLocal, self).__init__(options_page, parent)
-        self.options_page.init_regex_checker(self.ui.local_cover_regex_edit, self.ui.local_cover_regex_error)
-        self.ui.local_cover_regex_default.clicked.connect(self.set_local_cover_regex_default)
-
-    def set_local_cover_regex_default(self):
-        self.ui.local_cover_regex_edit.setText(self._DEFAULT_LOCAL_COVER_ART_REGEX)
-
-    def load(self):
-        self.ui.local_cover_regex_edit.setText(config.setting["local_cover_regex"])
-
-    def save(self):
-        config.setting["local_cover_regex"] = unicode(self.ui.local_cover_regex_edit.text())
-
-
 class ProviderTab(QtGui.QWidget):
 
     def __init__(self, provider, parent=None):
@@ -294,12 +106,9 @@ class CoverOptionsPage(OptionsPage):
             tab = ProviderTab(provider)
             self.ui.tab_cover_art_providers.addTab(tab, self.tabTitle(provider))
             tab.ui.enabled.setChecked(is_provider_enabled(provider.NAME))
-            # TODO : improve this
             options_widget = None
-            if provider.NAME == 'Cover Art Archive':
-                options_widget = ProviderOptionsCaa(self)
-            elif provider.NAME == 'Local':
-                options_widget = ProviderOptionsLocal(self)
+            if hasattr(provider, 'OPTIONS') and callable(provider.OPTIONS):
+                options_widget = provider.OPTIONS(self)
             if options_widget is not None:
                 tab.ui.scrollArea.setWidget(options_widget)
                 for method in ('load', 'save'):
